move load info to separate file
